% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_load_model.R
\name{wt_load_model}
\alias{wt_load_model}
\title{Load one of the implemented models and initialize}
\usage{
wt_load_model(model_name)
}
\arguments{
\item{model_name}{Character vector specifying the name of the model to be loaded.}
}
\description{
Loads one of the implemented models (so far only 'flair/ner-german-large') and initializes it.
}
\examples{
\dontrun{
wt_load_model(model_name = "ner_german_large")
}
}
