% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert_base_german_case_hatespeech.R
\name{wt_predict_bert_base_german_cased_hatespeech}
\alias{wt_predict_bert_base_german_cased_hatespeech}
\title{Perform Sequence Classification using deepset/bert-base-german-cased-hatespeech-GermEval18Coarse}
\usage{
wt_predict_bert_base_german_cased_hatespeech(sentence)
}
\arguments{
\item{sentence}{A character vector with a sentence compatible with the model.}
}
\description{
\code{wt_predict_labels_ner_german_large} returns a tidy data frame
of tokens classified as named entities including probabilities.
}
\examples{
\dontrun{
wt_load_model("ner_german_large")
# example sentence
sentence_example <- "Der Wichser muss raus."
# get predictions
wt_predict_bert_base_german_cased_hatespeech(sentence = sentence_example)
}
}
\seealso{
\url{https://huggingface.co/deepset/bert-base-german-cased-hatespeech-GermEval18Coarse} for
further information about this model.
}
