% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ner_german_large.R
\name{wt_predict_labels_ner_german_large}
\alias{wt_predict_labels_ner_german_large}
\title{Perform NER using flair/ner-german-large}
\usage{
wt_predict_labels_ner_german_large(sentence)
}
\arguments{
\item{sentence}{A character vector with a sentence compatible with the model.}
}
\description{
\code{wt_predict_labels_ner_german_large} returns a tidy data frame
of tokens classified as named entities including probabilities.
}
\examples{
\dontrun{
wt_load_model("ner_german_large")
# example sentence
sentence_example <- "Santesuisse und Curaviva fordern eine Umverteilung"
# get predictions
predict_labels_ner_german_large(sentence = sentence_example)
}
}
\seealso{
\url{https://huggingface.co/flair/ner-german-large} for
further information about this model.
}
