% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_load_model.R
\name{wt_load_model}
\alias{wt_load_model}
\title{Load one of the implemented models and initialize}
\usage{
wt_load_model(model_name)
}
\arguments{
\item{model_name}{Character vector specifying the name of the model (without author) to be loaded.}
}
\description{
Loads one of the following models:
\enumerate{
   \item flair/ner-german-large
   \item deepset/bert-base-german-cased-hatespeech-GermEval18Coarse
}
}
\examples{
\dontrun{
wt_load_model(model_name = "ner-german-large")
wt_load_model(model_name = "bert-base-german-cased-hatespeech-GermEval18Coarse")
}
}
